Fixes issue #112
