+++ load_confounds/parser.py
@@ -1,15 +1,10 @@
 """Flexible method to load confounds generated by fMRIprep.
 
-Authors: Hanad Sharmarke, Dr. Pierre Bellec, Francois Paugam
+Authors: load_confounds team
 """
 import numpy as np
 import pandas as pd
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import scale
-import warnings
-import os
-import json
-
+from . import confounds as cf
 
 # Global variables listing the admissible types of noise components
 all_confounds = [
@@ -21,25 +16,6 @@ all_confounds = [
     "ica_aroma",
     "censoring",
 ]
-
-
-def _add_suffix(params, model):
-    """
-    Add suffixes to a list of parameters.
-    Suffixes includes derivatives, power2 and full
-    """
-    params_full = params.copy()
-    suffix = {
-        "basic": {},
-        "derivatives": {"derivative1"},
-        "power2": {"power2"},
-        "full": {"derivative1", "power2", "derivative1_power2"},
-    }
-    for par in params:
-        for suff in suffix[model]:
-            params_full.append(f"{par}_{suff}")
-    return params_full
-
 
 def _check_params(confounds_raw, params):
     """Check that specified parameters can be found in the confounds."""
@@ -107,73 +83,6 @@ def _find_acompcor(confounds_json, n_compcor, acompcor_combined):
     return compcor_cols
 
 
-def _pca_motion(confounds_motion, n_components):
-    """Reduce the motion paramaters using PCA."""
-    n_available = confounds_motion.shape[1]
-    if n_components > n_available:
-        raise ValueError(
-            f"User requested n_motion={n_components} motion components, but found only {n_available}."
-        )
-    confounds_motion = confounds_motion.dropna()
-    confounds_motion_std = scale(
-        confounds_motion, axis=0, with_mean=True, with_std=True
-    )
-    pca = PCA(n_components=n_components)
-    motion_pca = pd.DataFrame(pca.fit_transform(confounds_motion_std))
-    motion_pca.columns = ["motion_pca_" + str(col + 1) for col in motion_pca.columns]
-    return motion_pca
-
-
-def _optimize_censoring(fd_outliers, n_scans):
-    """
-    Perform optimized censoring. After censoring volumes, further remove
-    continuous segments containing fewer than 5 volumes.
-    Power, Jonathan D., et al. "Methods to detect, characterize, and remove
-    motion artifact in resting state fMRI." Neuroimage 84 (2014): 320-341.
-    """
-    # Start by checking if the beginning continuous segment is fewer than 5 volumes
-    if fd_outliers[0] < 5:
-        fd_outliers = np.asarray(list(range(fd_outliers[0])) + list(fd_outliers))
-    # Do the same for the ending segment of scans
-    if n_scans - (fd_outliers[-1] + 1) < 5:
-        fd_outliers = np.asarray(
-            list(fd_outliers) + list(range(fd_outliers[-1], n_scans))
-        )
-    # Now do everything in between
-    fd_outlier_ind_diffs = np.diff(fd_outliers)
-    short_segments_inds = np.where(
-        np.logical_and(fd_outlier_ind_diffs > 1, fd_outlier_ind_diffs < 6)
-    )[0]
-    for ind in short_segments_inds:
-        fd_outliers = np.asarray(
-            list(fd_outliers) + list(range(fd_outliers[ind] + 1, fd_outliers[ind + 1]))
-        )
-    fd_outliers = np.sort(np.unique(fd_outliers))
-    return fd_outliers
-
-
-def _confounds_to_df(confounds_raw):
-    """Load raw confounds as a pandas DataFrame."""
-    if "nii" in confounds_raw[-6:]:
-        suffix = "_space-" + confounds_raw.split("space-")[1]
-        confounds_raw = confounds_raw.replace(suffix, "_desc-confounds_timeseries.tsv",)
-        # fmriprep has changed the file suffix between v20.1.1 and v20.2.0 with respect to BEP 012.
-        # cf. https://neurostars.org/t/naming-change-confounds-regressors-to-confounds-timeseries/17637
-        # Check file with new naming scheme exists or replace, for backward compatibility.
-        if not os.path.exists(confounds_raw):
-            confounds_raw = confounds_raw.replace(
-                "_desc-confounds_timeseries.tsv", "_desc-confounds_regressors.tsv",
-            )
-
-    # Load JSON file
-    with open(confounds_raw.replace("tsv", "json"), "rb") as f:
-        confounds_json = json.load(f)
-
-    confounds_raw = pd.read_csv(confounds_raw, delimiter="\t", encoding="utf-8")
-
-    return confounds_raw, confounds_json
-
-
 def _sanitize_confounds(confounds_raw):
     """Make sure the inputs are in the correct format."""
     # we want to support loading a single set of confounds, instead of a list
@@ -196,25 +105,6 @@ def _sanitize_strategy(strategy):
     else:
         raise ValueError("strategy needs to be a list of strings")
     return strategy
-
-
-def _confounds_to_ndarray(confounds, demean):
-    """Convert confounds from a pandas dataframe to a numpy array."""
-    # Convert from DataFrame to numpy ndarray
-    labels = confounds.columns
-    confounds = confounds.values
-
-    # Derivatives have NaN on the first row
-    # Replace them by estimates at second time point,
-    # otherwise nilearn will crash.
-    mask_nan = np.isnan(confounds[0, :])
-    confounds[0, mask_nan] = confounds[1, mask_nan]
-
-    # Optionally demean confounds
-    if demean:
-        confounds = scale(confounds, axis=0, with_std=False)
-
-    return confounds, labels
 
 
 def _raise_conf_not_found_error(not_found_conf, not_found_keys):
@@ -413,7 +303,7 @@ class Confounds:
     def _load_single(self, confounds_raw):
         """Load a single confounds file from fmriprep."""
         # Convert tsv file to pandas dataframe
-        confounds_raw, self.json_ = _confounds_to_df(confounds_raw)
+        confounds_raw, self.json_ = cf._confounds_to_df(confounds_raw)
 
         confounds = pd.DataFrame()
         not_found_conf = []
@@ -430,12 +320,12 @@ class Confounds:
         if not_found_conf or not_found_keys:
             _raise_conf_not_found_error(not_found_conf, not_found_keys)
 
-        confounds, labels = _confounds_to_ndarray(confounds, self.demean)
+        confounds, labels = cf._confounds_to_ndarray(confounds, self.demean)
         return confounds, labels
 
     def _load_motion(self, confounds_raw):
         """Load the motion regressors."""
-        motion_params = _add_suffix(
+        motion_params = cf._add_suffix(
             ["trans_x", "trans_y", "trans_z", "rot_x", "rot_y", "rot_z"], self.motion
         )
         _check_params(confounds_raw, motion_params)
@@ -443,7 +333,7 @@ class Confounds:
 
         # Optionally apply PCA reduction
         if self.n_motion > 0:
-            confounds_motion = _pca_motion(confounds_motion, n_components=self.n_motion)
+            confounds_motion = cf._pca_motion(confounds_motion, n_components=self.n_motion)
         return confounds_motion
 
     def _load_high_pass(self, confounds_raw):
@@ -453,13 +343,13 @@ class Confounds:
 
     def _load_wm_csf(self, confounds_raw):
         """Load the regressors derived from the white matter and CSF masks."""
-        wm_csf_params = _add_suffix(["csf", "white_matter"], self.wm_csf)
+        wm_csf_params = cf._add_suffix(["csf", "white_matter"], self.wm_csf)
         _check_params(confounds_raw, wm_csf_params)
         return confounds_raw[wm_csf_params]
 
     def _load_global(self, confounds_raw):
         """Load the regressors derived from the global signal."""
-        global_params = _add_suffix(["global_signal"], self.global_signal)
+        global_params = cf._add_suffix(["global_signal"], self.global_signal)
         _check_params(confounds_raw, global_params)
         return confounds_raw[global_params]
 
@@ -505,7 +395,7 @@ class Confounds:
         )
         # Do optimized scrubbing if desired
         if self.censoring == "optimized":
-            combined_outliers = _optimize_censoring(combined_outliers, n_scans)
+            combined_outliers = cf._optimize_censoring(combined_outliers, n_scans)
         # Make one-hot encoded motion outlier regressors
         motion_outlier_regressors = pd.DataFrame(
             np.transpose(np.eye(n_scans)[combined_outliers]).astype(int)
